Split `Importer` and `Importable` into separate modules
