+++ src/viur/toolkit/importer/__init__.py
@@ -1,0 +1,6 @@
+"""
+Tools for copying data from ViUR 1.x to ViUR 3.x systems by using the JSON interfaces.
+"""
+
+from .importable import Importable
+from .importer import Importer
